Call weather API
Calls the weather API with the long and lat coordinates returned from the hook added in the previous commit.

The useEffect hook takes shouldDisplayWeather as a dependency so that it is called when the longitude and latitude fields update with the correct coordinates returned from the useCoordinates hook.

The shouldDisplayWeather is set to true when both longitude and latitude have been updated from their initial value of 0.

The weather API call is saved to the state in an array called ‘result’ composed of objects of the Value interface.

// eslint-disable-next-line is used to avoid passing in coordinate.lat and coordinate.long to the dependency array of the useEffect and causing more re-renders than we need.